Add files via upload
Exploratory Data Analysis¶
Description of EDA:
EDA, which stands for Exploratory Data Analysis, refers to the process of analyzing and summarizing data sets in order to gain insights, discover patterns, and identify relationships or anomalies within the data. It is a crucial step in the data analysis pipeline and helps to understand the underlying structure and characteristics of the data.

During EDA, various statistical techniques, visualizations, and data manipulation methods are applied to explore the data. The goals of EDA include:

Understanding the data: EDA aims to gain a deep understanding of the data by examining its distribution, range, and key statistics such as mean, median, and standard deviation. It involves assessing the quality, completeness, and consistency of the data.

Detecting patterns and relationships: EDA helps to identify patterns, trends, or relationships between variables within the data. This can involve exploring correlations, dependencies, or associations using techniques such as scatter plots, heatmaps, and correlation matrices.

Handling missing or inconsistent data: EDA involves examining missing values, outliers, or inconsistent data points. It may include strategies such as imputation of missing values, outlier detection, or data cleansing to ensure the data quality is suitable for further analysis.

Feature engineering and selection: EDA helps in identifying relevant features or variables that have a significant impact on the target variable. It involves assessing the importance and relevance of features, identifying redundant or irrelevant ones, and creating new features if necessary.

Visualization and communication: EDA often involves creating visualizations such as histograms, box plots, bar charts, or interactive dashboards to present the data visually. These visual representations help in understanding the data and effectively communicating the findings to stakeholders.

Overall, EDA plays a crucial role in the data analysis process, providing a solid foundation for subsequent modeling, hypothesis testing, and decision-making. It helps in gaining insights, formulating hypotheses, and generating meaningful questions to guide further analysis and exploration of the data